% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{add_averted_columns}
\alias{add_averted_columns}
\title{calculate deaths and overdoses averted}
\usage{
add_averted_columns(model_data)
}
\arguments{
\item{model_data}{A data frame containing model output. Must include columns:
\itemize{
\item \code{date}: a Date object,
\item \code{year_quarter}: a grouping variable (e.g., "2021 Q2"),
\item \code{model}: model identifier,
\item \code{run}: simulation run identifier,
along with numeric variables to be summarized.
}}
}
\value{
tibble
}
\description{
calculate deaths and overdoses averted
}
