% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plots.R
\name{get_low_medium_high_breaks}
\alias{get_low_medium_high_breaks}
\title{get result breaks into small, medium, and large categories}
\usage{
get_low_medium_high_breaks(result)
}
\arguments{
\item{result}{output of \code{\link[=load_sensitivity_analysis_data]{load_sensitivity_analysis_data()}}}
}
\value{
vector of length 4 representing the 0th 33rd, 66th and 100th
percentiles
}
\description{
get result breaks into small, medium, and large categories
}
