% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plots.R
\name{get_max_min_breaks}
\alias{get_max_min_breaks}
\title{get max and min deaths_averted from results}
\usage{
get_max_min_breaks(result)
}
\arguments{
\item{result}{output of \code{\link[=load_sensitivity_analysis_data]{load_sensitivity_analysis_data()}}}
}
\value{
vector of length 2 representing min and max
}
\description{
get max and min deaths_averted from results
}
