% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plots.R
\name{plot_model_difference_heatmap}
\alias{plot_model_difference_heatmap}
\title{Plot Model Difference Heatmap}
\usage{
plot_model_difference_heatmap(result, sensitivity_analysis = "oat_ret")
}
\arguments{
\item{result}{A list containing OAT results with \code{x_label}, \code{y_label},
\code{x_tick_labels}, \code{y_tick_labels}, and \code{data}.}

\item{sensitivity_analysis}{Optional name of the element within \code{result} containing OAT results.
Can be either \code{"pop_mort"} or \code{"oat_ret"}
Defaults to \code{"oat_ret"}.}
}
\value{
A patchwork object combining the main heatmap and custom legend.
}
\description{
Creates a bivariate heatmap comparing the difference in total deaths between
two models ("OTEM" and "BCROM"), including a custom legend laid out as a 3×3 grid.
}
