% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summarise_deaths_averted}
\alias{summarise_deaths_averted}
\title{summarise deaths averted}
\usage{
summarise_deaths_averted(model_data, digits = 2)
}
\arguments{
\item{model_data}{output of \code{\link[=load_data]{load_data()}}}

\item{digits}{number of significant digits}
}
\value{
\code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
summarise deaths averted
}
