% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summarise_events_averted}
\alias{summarise_events_averted}
\title{summarise deaths averted}
\usage{
summarise_events_averted(model_data, event = "deaths", digits = 2)
}
\arguments{
\item{model_data}{output of \code{\link[=load_data]{load_data()}}}

\item{event}{Either \code{"deaths"} or \code{"overdoses"}. Default is \code{"deaths"}}

\item{digits}{number of significant digits}
}
\value{
\code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
summarise deaths averted
}
